backend/socket_server/requirements.txt
python-socketio
redis
eventlet
google-generativeai
